+++ CHANGELOG.md
@@ -10,6 +10,8 @@ CHANGELOG
 
 - Add missing builtin `MapArray` to Go SDK.
   [#4144](https://github.com/pulumi/pulumi/pull/4144)
+- Add aliases to Go SDK codegen pkg.
+  [#4157](https://github.com/pulumi/pulumi/pull/4157)
 
 - Discontinue testing on Node 8 (which has been end-of-life since January 2020), and start testing on Node 13.
   [#4156](https://github.com/pulumi/pulumi/pull/4156)